requirements.txt
# Core framework
crewai>=0.80.0
crewai-tools>=0.14.0

# LLM & Embeddings
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
sentence-transformers>=3.0.0

# Vector store
chromadb>=0.5.0

# Re-ranking
torch>=2.0.0

# LLM Provider (required by CrewAI for Groq)
litellm>=1.0.0

# UI
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
